Commit 11/03/2023

+++ app.py
@@ -9,8 +9,8 @@ app = Flask(__name__)
 app.config.from_object(DevelopmentConfig)
 csrf = CSRFProtect()
 
-@app.route("/", methods = ['GET', 'POST'])
-def index():
+@app.route("/insertar", methods = ['GET', 'POST'])
+def insertar():
     reg_alumno = forms.UserForm(request.form)
     if request.method == 'POST':
         alumno = Alumnos(
@@ -21,9 +21,11 @@ def index():
 
         db.session.add(alumno)
         db.session.commit()
+        return redirect(url_for('ABCompleto'))
+    
     return render_template('index.html', form = reg_alumno)
 
-@app.route("/ABCompleto", methods = ['GET', 'POST'])
+@app.route("/", methods = ['GET', 'POST'])
 def ABCompleto():
     reg_aluno = forms.UserForm(request.form)
     # SELECT * FROM alumnos
@@ -58,23 +60,29 @@ def modificar():
     
     return render_template('modificar.html', form = reg_alumno)
 
-@app.route("/eliminar", methods = ['GET'])
+@app.route("/eliminar", methods = ['GET', 'POST'])
 def eliminar():
     reg_alumno = forms.UserForm(request.form)
 
-    id = request.args.get('id')
+    if request.method == 'GET':
+        id = request.args.get('id')
+        #SELECT * FROM alumnos WHERE id = id
+        alumno = db.session.query(Alumnos).filter(Alumnos.id == id).first()
 
-    alumno = db.session.query(Alumnos).filter(Alumnos.id == id).first()
+        reg_alumno.id.data = alumno.id
+        reg_alumno.nombre.data = alumno.nombre
+        reg_alumno.apellidos.data = alumno.apellidos
+        reg_alumno.email.data = alumno.email
 
-    alumno.nombre = reg_alumno.nombre.data
-    alumno.apellidos = reg_alumno.apellidos.data
-    alumno.email = reg_alumno.email.data
+    if request.method == 'POST':
+        id = reg_alumno.id.data
+        alumno = db.session.query(Alumnos).filter(Alumnos.id == id).first()
 
-    db.session.delete(alumno)
-    db.session.commit()
+        db.session.delete(alumno)
+        db.session.commit()
+        return redirect(url_for('ABCompleto'))
     
-    return redirect(url_for('ABCompleto'))
-    
+    return render_template('eliminar.html', form = reg_alumno)
 
 if __name__ == "__main__":
     csrf.init_app(app)
